Refactor server helpers into reusable modules
