+++ Recipe/src/lib/date.js
@@ -1,0 +1,10 @@
+function toIsoDate(value) {
+  if (!value) return '';
+  const d = value instanceof Date ? value : new Date(value);
+  if (Number.isNaN(d.getTime())) return '';
+  return d.toISOString().split('T')[0];
+}
+
+module.exports = {
+  toIsoDate
+};
