+++ Recipe/src/lib/normaliseError.js
@@ -1,0 +1,42 @@
+const ValidationError = require('../errors/ValidationError');
+
+function normaliseError(err) {
+  if (!err) return err;
+
+  if (err instanceof ValidationError) {
+    return err;
+  }
+
+  if (err.name === 'ValidationError' && err.errors) {
+    const messages = [];
+    for (const key in err.errors) {
+      if (Object.prototype.hasOwnProperty.call(err.errors, key)) {
+        messages.push(err.errors[key].message);
+      }
+    }
+    return new ValidationError(messages);
+  }
+
+  if (err.code === 11000) {
+    const details = [];
+    if (err.keyValue) {
+      for (const key in err.keyValue) {
+        if (Object.prototype.hasOwnProperty.call(err.keyValue, key)) {
+          details.push(key + ' already exists');
+        }
+      }
+    }
+    if (!details.length) {
+      details.push('Duplicate value is not allowed');
+    }
+    return new ValidationError(details);
+  }
+
+  if (err.message && err.message.indexOf('Quantity cannot be negative') !== -1) {
+    return new ValidationError([err.message]);
+  }
+
+  return err;
+}
+
+module.exports = normaliseError;
